Colocate 721delegate hooks
