+++ src/components/Project/ProjectHeader/ProjectHeader.tsx
@@ -7,8 +7,8 @@ import { GnosisSafeBadge } from 'components/Project/ProjectHeader/GnosisSafeBadg
 import ProjectLogo from 'components/ProjectLogo'
 import { ProjectMetadataContext } from 'contexts/projectMetadataContext'
 import { ThemeContext } from 'contexts/themeContext'
-import { useGnosisSafe } from 'hooks/GnosisSafe'
 import useMobile from 'hooks/Mobile'
+import { useGnosisSafe } from 'hooks/safe/GnosisSafe'
 import { useContext } from 'react'
 import { EditProjectHandleButton } from './EditProjectHandleButton'
 import SocialLinks from './SocialLinks'
