+++ src/components/ProjectSafeDashboard/ProjectSafeDashboard.tsx
@@ -5,7 +5,7 @@ import Loading from 'components/Loading'
 import { Tab } from 'components/Tab'
 import { layouts } from 'constants/styles/layouts'
 import { ThemeContext } from 'contexts/themeContext'
-import { useGnosisSafe } from 'hooks/GnosisSafe'
+import { useGnosisSafe } from 'hooks/safe/GnosisSafe'
 import { useQueuedSafeTransactions } from 'hooks/safe/QueuedSafeTransactions'
 import { generateSafeUrl } from 'lib/safe'
 import { SafeTransactionType } from 'models/safe'
