+++ src/hooks/JB721Delegate/contractReader/NftCollectionSymbol.ts
@@ -1,6 +1,6 @@
 import { useJB721TieredDelegate } from 'hooks/contracts/JB721Delegate/useJB721TieredDelegate'
 
-import useV2ContractReader from './V2ContractReader'
+import useV2ContractReader from '../../v2v3/contractReader/V2ContractReader'
 
 export function useNftCollectionSymbol(dataSourceAddress: string | undefined) {
   const delegateContract = useJB721TieredDelegate({
