+++ src/hooks/JB721Delegate/contractReader/NftRewardContractUri.ts
@@ -1,16 +1,9 @@
 import * as constants from '@ethersproject/constants'
 import { useStoreOfJB721TieredDelegate } from 'hooks/contracts/JB721Delegate/useStoreofJB721TieredDelegate'
-import { JB721TierParams } from 'models/nftRewardTier'
-import { MAX_NFT_REWARD_TIERS } from 'utils/nftRewards'
-import useV2ContractReader from './V2ContractReader'
+import useV2ContractReader from '../../v2v3/contractReader/V2ContractReader'
 
-export function useNftRewardTiersOf({
-  dataSourceAddress,
-  limit,
-}: {
-  dataSourceAddress: string | undefined
-  limit?: number
-}) {
+// get the contract URI (cid for collection's metadata on IPFS) of an NFT collection
+export function useNftRewardContractUri(dataSourceAddress: string | undefined) {
   const JBTiered721DelegateStore = useStoreOfJB721TieredDelegate({
     JB721TieredDelegateAddress: dataSourceAddress,
   })
@@ -18,12 +11,10 @@ export function useNftRewardTiersOf({
   const hasDataSource =
     dataSourceAddress && dataSourceAddress !== constants.AddressZero
 
-  return useV2ContractReader<JB721TierParams[]>({
+  return useV2ContractReader<string>({
     contract: JBTiered721DelegateStore,
-    functionName: 'tiers',
+    functionName: 'contractUriOf',
     // send null when project has no dataSource, so the fetch doesn't execute.
-    args: hasDataSource
-      ? [dataSourceAddress, 0, limit ?? MAX_NFT_REWARD_TIERS]
-      : null,
+    args: hasDataSource ? [dataSourceAddress] : null,
   })
 }
