+++ src/hooks/JB721Delegate/contractReader/NftRewardTiersOf.ts
@@ -1,9 +1,16 @@
 import * as constants from '@ethersproject/constants'
 import { useStoreOfJB721TieredDelegate } from 'hooks/contracts/JB721Delegate/useStoreofJB721TieredDelegate'
-import useV2ContractReader from './V2ContractReader'
+import { JB721TierParams } from 'models/nftRewardTier'
+import { MAX_NFT_REWARD_TIERS } from 'utils/nftRewards'
+import useV2ContractReader from '../../v2v3/contractReader/V2ContractReader'
 
-// get the contract URI (cid for collection's metadata on IPFS) of an NFT collection
-export function useNftRewardContractUri(dataSourceAddress: string | undefined) {
+export function useNftRewardTiersOf({
+  dataSourceAddress,
+  limit,
+}: {
+  dataSourceAddress: string | undefined
+  limit?: number
+}) {
   const JBTiered721DelegateStore = useStoreOfJB721TieredDelegate({
     JB721TieredDelegateAddress: dataSourceAddress,
   })
@@ -11,10 +18,12 @@ export function useNftRewardContractUri(dataSourceAddress: string | undefined) {
   const hasDataSource =
     dataSourceAddress && dataSourceAddress !== constants.AddressZero
 
-  return useV2ContractReader<string>({
+  return useV2ContractReader<JB721TierParams[]>({
     contract: JBTiered721DelegateStore,
-    functionName: 'contractUriOf',
+    functionName: 'tiers',
     // send null when project has no dataSource, so the fetch doesn't execute.
-    args: hasDataSource ? [dataSourceAddress] : null,
+    args: hasDataSource
+      ? [dataSourceAddress, 0, limit ?? MAX_NFT_REWARD_TIERS]
+      : null,
   })
 }
