+++ src/providers/v2v3/NftRewardsProvider.tsx
@@ -1,9 +1,9 @@
 import { NftRewardsContext } from 'contexts/nftRewardsContext'
 import { ProjectMetadataContext } from 'contexts/projectMetadataContext'
 import { V2V3ProjectContext } from 'contexts/v2v3/V2V3ProjectContext'
+import { useNftCollectionMetadataUri } from 'hooks/JB721Delegate/contractReader/NftCollectionMetadataUri'
+import { useNftRewardTiersOf } from 'hooks/JB721Delegate/contractReader/NftRewardTiersOf'
 import useNftRewards from 'hooks/NftRewards'
-import { useNftCollectionMetadataUri } from 'hooks/v2v3/contractReader/NftCollectionMetadataUri'
-import { useNftRewardTiersOf } from 'hooks/v2v3/contractReader/NftRewardTiersOf'
 import { useContext } from 'react'
 import { EMPTY_NFT_COLLECTION_METADATA } from 'redux/slices/editingV2Project'
 import { CIDsOfNftRewardTiersResponse } from 'utils/nftRewards'
